Fixes #39735 and 40257
2003-04-15  JP Rosevear  <jpr@ximian.com>

	Fixes #39735 and 40257

	* gui/gnome-cal.c: convert float pane positions to ints
	(setup_widgets): set the initial position after realization and
	track the drags to get the new position, pack the panes slightly
	differently
	(gnome_calendar_set_pane_positions): set purely pixel oriented
	positions
	(gnome_calendar_update_config_settings): no need to update quanta
	setting
	(gnome_calendar_hpane_realized): realization callback
	(gnome_calendar_vpane_realized): ditto
	(gnome_calendar_vpane_resized): resize callback, store new size
	(gnome_calendar_hpane_resized): ditto

	* gui/calendar-model.c (get_due_status): handle an error getting
	the timezone

	* gui/calendar-config.c (calendar_config_get_hpane_pos): return an
int
	(calendar_config_set_hpane_pos): take an int
	(calendar_config_get_vpane_pos): return an int
	(calendar_config_set_vpane_pos): take an int

	* gui/calendar-config.h: update protos

	* gui/apps_evolution_calendar.schemas: update defaults for pane
	positions

svn path=/trunk/; revision=20863

+++ calendar/gui/calendar-config.h
@@ -117,17 +117,17 @@ gint	  calendar_config_get_default_view	(void);
 void	  calendar_config_set_default_view	(gint	      view);
 
 /* The positions of the panes in the normal and month views. */
-gfloat    calendar_config_get_hpane_pos		(void);
-void	  calendar_config_set_hpane_pos		(gfloat	      hpane_pos);
+gint      calendar_config_get_hpane_pos		(void);
+void	  calendar_config_set_hpane_pos		(gint	      hpane_pos);
 
-gfloat    calendar_config_get_vpane_pos		(void);
-void	  calendar_config_set_vpane_pos		(gfloat	      vpane_pos);
+gint      calendar_config_get_vpane_pos		(void);
+void	  calendar_config_set_vpane_pos		(gint	      vpane_pos);
 
-gfloat    calendar_config_get_month_hpane_pos	(void);
-void	  calendar_config_set_month_hpane_pos	(gfloat	      hpane_pos);
+gint      calendar_config_get_month_hpane_pos	(void);
+void	  calendar_config_set_month_hpane_pos	(gint	      hpane_pos);
 
-gfloat    calendar_config_get_month_vpane_pos	(void);
-void	  calendar_config_set_month_vpane_pos	(gfloat	      vpane_pos);
+gint      calendar_config_get_month_vpane_pos	(void);
+void	  calendar_config_set_month_vpane_pos	(gint	      vpane_pos);
 
 /* Colors for the task list */
 const char *calendar_config_get_tasks_due_today_color	(void);
